+++ tables-kt/tables-kt-dataframe/src/test/kotlin/space/kscience/dataforge/dataframe/DataFrameTableTest.kt
@@ -2,10 +2,13 @@ package space.kscience.dataforge.dataframe
 
 import org.jetbrains.kotlinx.dataframe.api.add
 import org.junit.jupiter.api.Test
+import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.misc.DFExperimental
 import space.kscience.tables.*
 import kotlin.math.pow
+import kotlin.reflect.typeOf
 import kotlin.test.assertEquals
+import kotlin.test.assertFails
 import kotlin.test.assertTrue
 
 @OptIn(DFExperimental::class)
@@ -35,6 +38,27 @@ internal class DataFrameTableTest {
         assertEquals(table.columns[x], newTable.columns[x])
         assertTrue {
             table.rowsToColumn("z") { it[x] + it[y] + 1.0 }.contentEquals(newTable.columns["z"])
+        }
+    }
+
+    @Test
+    fun testDataFrameAccessors() {
+        val x by ColumnHeader.typed<Double>()
+        val y by ColumnHeader.typed<Number>()
+        val y2 = SimpleColumnHeader<Int>("y", typeOf<Int>(), Meta.EMPTY)
+
+        val table = ColumnTable<Number?>(100) {
+            //filling column with double values equal to index
+            fill(x) { it.toDouble() }
+            fill(y) { it.toDouble() }
+        }
+
+        val dataFrame = table.toDataFrame()
+
+        assertEquals(table[2, x], dataFrame[2][x])
+        assertEquals(table[2, y], dataFrame[2][y])
+        assertFails {
+            dataFrame[2][y2]
         }
     }
 }